Add interfaces for retrieving separate linkable downstream binaries (#5128)
* Implement separate downstream library interface

Create a new com interface to house the methods for
precompiling slang modules to target code.

Add methods to count dependent modules and scrape
them for downstream target binaries such that the
downstream target binaries are linkabe outside
of slang, e.g. via spirv-link or dxc.

Fixes #5147

* Rename to _Experimental

Clearly identify this as an interface subject to change.