Enable Link-Time-Code-Generation for Release builds (#5678)
This commit enables Link Time Code Generation, or LTCG for short, for
the release builds.

Because "incremental" build is enabled, this ends up using
"/LTCG:incremental" rather than the full "/LTCG". It is still expected
to more performant than without any "/LTCG" at all.

In order to use the full LTCG, we will need to give up the incremental
build option for the Release build. We will need to decide if we wants
to do it later.