Update Slang-RHI/slang (#5701)
* Update Slang-RHI/slang

This brings in new fixes for WebGPU.
In particular, the "use_dxc" toggle is now used, which should enable these tests to run on
WebGPU, if f16 is otherwise supported:

- `tests/language-feature/generics/variadic-0.slang`
- `tests/language-feature/generics/tuple.slang`

This closes #5605.

* Disable tests/autodiff/float-cast.slang for wgpu

This test was previously not running for WebGPU because it required the 'half' render
feature, and Slang-RHI was previously not reporting it.

With the Slang-RHI update, the test now runs on WebGPU.
It now fails because the test is using 'double' which is just not supported on WebGPU.

Thus this commit disables the test.