Core lib Metal math function fixes
In order to emit fast target implementations of some Metal-based
functions (fmin(), fmax(), fmin3(), fmax3(), fmedian3()) on all
targets, remove some specification regarding the handling of NaNs,
and also remove the enforcement of the specification.

These functions are now documented to be basically undefined now
in the presence of NaN input, to make the common "is a number"
case fast.

Also, clarify that powr() is undefined when given a non-positive
base input value, allowing us to remove an additional abs()
operation that was unnecessarily coercing results to be predictable
on non-Metal targets.

Closes #5580
Closes #5581
Closes #5587