CI: Build the full slang-wasm.[js|wasm] targets instead of just the slang library.
A recent build breakage for slang-wasm.[js|wasm] was not noticed by CI since it built the
slang library target instead of the slang-wasm.js "executable" target.
(We added a that used objcopy to split debug info from executables. objcopy does not
report errors when it finds object files with unexpected format inside static libraries,
but it *does* report errors when it's run on an exceutable of unexpected format, such as
in the case of slang-wasm.js.)

This closes #5959.