slang-test: Ensure runtime is consistent (#6021)
I hit issue #5305 again after re-generating the build directory, building and then running
slang-test.
In other words, D3D12 and Slang-RHI/WebGPU/D3D12 failed to create a device due to the
runtime being incosistent. Bilding copy-prebuilt-binaries fixed it, and so it's best to
make this step a dependency of slang-test.