Upgrade actions/upload-artifact from v3 to v4
The version 3 of actions/upload-artifact will be deprecated soon.
https://github.blog/changelog/2024-04-16-deprecation-notice-v3-of-the-artifact-actions/

This commit upgrades to v4.
There is a document describing how to migrate/upgrade:
https://github.com/actions/upload-artifact/blob/main/docs/MIGRATION.md

The breaking change on v4 is that the uploaded files are immutable and
we can no longer append files with the same name. We don't seem to be
doing it so it should be fine for us.

Closes https://github.com/shader-slang/slang/issues/6050