Fix potential test failures due to SPIRV validation failure
There are a few tests that are currently passing but could fail when
conditions change little bit.

When slang-test runs with `test-server`, the spirv validation fails and
the test gets reported as failed even though they are actually passing.

To avoid the potential problem, this commit adds an empty entry point.