Fix test errors
In ResolveInvoke() call, we add a special case for one-argument
Invoke resolution. Because slang will treat this as a type conversion

Type1 A;
Type2 A = Type2(B);

We we synthesize the constructor invocation, we have to bypass this
special handling, because this invocation is converted from initialize
list, so want to call the constructor directly.