Fix autodiff issue for vector<T, N> (#6275)
* Fix autodiff issue for vector<T, N>

Close #6154

We didn't implement correctly for vector<T, N> regarding the
differentiablity. As we check differentiable before specialization,
however according to the definition of vector, it has to be specialized
to IFloat to know it's conformed to IDifferential type. Therefore for
parameter type vector<T, N> will become no_diff.

Therefore, we change the implementation a to make it explicit conform
to IDifferential type.

* fix typo