Deduplicate witness table based on type and operand-0
This commit is to experiment the idea of what would happen if we
deduplicate the witness table based only on its type and its first
operand.

It turned out that it causes a crash, because it doesn't carry the
decorations and childs properly. It means we need to also consider them
as well as its type and its first operand. But if we do, we cannot
deduplicate anything.

It seems that this is not the right place to apply the deduplication.