Fix build error when SLANG_ENABLE_TESTS is set to OFF
When SLANG_ENABLE_TESTS is disabled but SLANG_ENABLE_GFX is enabled,
cmake was failing. It turned out that gfx build target was trying to
copy some of files to the directory where slang-test uses. When
SLANG_ENABLE_TESTS is disabled, the directory information became
unavailable, because slang-test target became unavailable.

This commit makes the copying behavior conditional on SLANG_ENABLE_TESTS
in the gfx build target.