Fix SPIRV error for WorkGraphs
This commit resolves an error message that is caused while emitting
SPIRV code.

The cause was on the fact that there is no corresponding IRInst for
one of SPIRV instructions when emitting,
```
%in_payload_t = OpTypeNodePayloadArrayAMDX $$T;
%in_payload_ptr_t = OpTypePointer NodePayloadAMDX %in_payload_t;
```

There is an IRInst for `OpTypePointer` but there isn't IRInst
corresponding to `OpTypeNodePayloadArrayAMDX`. When Slang emits
`OpTypePointer` with its second operand `OpTypeNodePayloadArrayAMDX`,
it is going to be emitted without a matching IRInst.