Sharpen condition on warning about used capabilities relative to selected profile
If the profile/capability in options set, but set to unknown, then don't count the
setting as requesting a particular profile/capability.

This helps in addressing issue #4760 because the new session-based compilation API always
adds a profile setting via Linkage::addTarget, even if the profile is unknown.

That caused lots of tests to fail due to unexpected output:
"warning 41012: entry point 'computeMain' uses additional capabilities that are not part of the specified profile..."

The old compilation request -based API did not add profile/capabiltiy to the options, and
so did not generate this warning.