Add module reference options to options list when parsing them (#6539)
* Add module reference options to options list when parsing them

Previously, the module was just added to the compile request directly.
This is a problem because the non-deprecated command line option parsing interface
IGlobalSession::parseCommandLineArguments just creates a temporary compile request.
Therefore these options are currently not detected by the API.

This helps to address issue #4760 since it will need to access these modules in order to
supply all of the required code.

* format code

---------

Co-authored-by: slangbot <186143334+slangbot@users.noreply.github.com>