Add referenced modules as libraries when creating a session (#6569)
In the legacy compile request based API, the referenced modules are added to the request's
linkage libraries as part of compiler option parsing.
In the non-legacy compilation API, the argument parsing creates a temprary compile request
and so those libraries only survive as options.

This change will look for such options when creating an ISession object, and again add the
referenced modules to the libraries of the new linkage that's contained in the ISession
object.

This is done in two steps:

1. Factor out a helper to create a referenced module artifact in the same way as it's done
   during legacy option parsing.
2. Use the helper function to create artifacts to add to the linkage libararies, when the
   session is created.

This helps to address issue #4760, because it enables passing in downstream modules via
options, as is required for the following tests:

tests/library/library-test.slang.2 (dx12)
tests/library/export-test.slang.2 (dx12)