Fix issue in peepholeOptimize (#6543)
Close #6541.

Previously in type legalization pass, we skip the VoidType field when call make_struct, however in some optimization pass we keep counting the VoidType field. We have to make this behavior consistently over all our codebase.

So in this change, we spot the make_struct call and leave VoidType field as it.