Add support for Array Sizes using Generic arguments to be initialized via {}
Fixes one subissue of #6138

This change adds support for initializing Arrays with Generic size arguments via {}
and adds a test to verify it.

The change checks for an array whose size parameter is a GenericParamIntVal
and since the size of such an array will be known at link time, is not considered
as a case of the size not being known statically.