Implement 64bit countbits intrinsic (#6433)
Change modifies the countbits intrinsic to use generics in order to
support 64bit countbits on select platforms where this is supported.
On platforms where this is not natively supported, we emulate by
converting the 64-bit type into a uint2 (metal and spir-v).

This should align with the implementation of other uint64_t
intrinsics such as abs, min, max and clamp.

Added new countbits64 test to verify changes.

Updated documentation for 64bit-type-support.html