Add unit test
Add unit test.

Also add check for the function parameter when we create existential
type for function parameter. Because if the argument is not ready to be
extracted from makeExistential, but the existential is already fully
specialized, we could replace the function parameter with the concrete
type too early. In this case, we will also lose the change to specialize
the callee.