Remove support for "module libraries"
This change *also* removes support for the `-r` command-line option to
`slangc`, and the ability to pass `.slang-module` files as
command-line arguments.

It may turn out that some (or all) of the functionality being removed
in this change is required by at least some existing users of the
slang API or command-line compiler. The intention of this change is
to provide a way to evaluate whether these changes would break any
user code that is being tested by various automated CI processes.